Export the Plugin (entrypoint) directly
This will bring alignment with other Markdown-It plugins that don't require the developer to call .markdownPlugin when adding to Markdown-It. More so, it allows for looping of plugins and dynamically loading into Markdown-It.

All exports are still available via their internal module exports.

Added a new test to demostate the typical integration code. Note: a workaround is needed owing to an outstanding issue where <p> tags are wrapping the expected output.